🚧 progress: prop 오류 수정 작업 진행 중
